// schemas/user.ts
export const updateUserBaseSchema = z.object({
  name: z.string().min(1).optional(),
  email: z.string().email().optional(),
  system: z.boolean().optional(),
  roleId: z.string().optional(),

  //   password: z.string().min(6).optional(),
});

export const updateUserPreferencesSchema = z.object({
  theme: z.enum(["light", "dark"]).optional(),
  communication: z.array(z.string()).optional(),
  language: z.string().optional(),
  currency: z.string().optional(),
});

export const updateUserImageSchema = z.object({
  imageId: z.string(),
  data: z.object({
    name: z.string().optional(),
    private: z.boolean().optional(),
  }),
});